All AsyncStorage fn from Login in file, GetEvents is auth
