+++ MobileUI/API/CalendarAPIHandling.js
@@ -1,13 +1,16 @@
 import React, {useContext} from 'react';
-import UserContext from '../contexts/User';
+// import AsyncStorage from '@react-native-async-storage/async-storage';
+import {getUsernameValue} from '../miscHelpers/AsyncStorageMethods';
 
 /**
  * API call to get calendar events
  * FYI: under postmanTest1, capstone project calendar id = 7
  * @returns result of GetEvents API request as a JSON object
  */
-export async function calendarGetEvents() {
+export async function calendarGetEvents(name) {
+  // const user = useContext(UserContext);
   console.log('(CAPIHandling) Beginning of CalendarGetEvents');
+  let user = await getUsernameValue(name);
   try {
     // console.log('Sending Username: ' + name + ' Password: ' + password);
     // So what needs to change is we need to send name and pass, and get the 'token' and 'expiration'.
@@ -15,6 +18,9 @@ export async function calendarGetEvents() {
       'http://ec2-3-84-219-120.compute-1.amazonaws.com/EventModels/GetEvents',
       {
         method: 'GET',
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       },
     );
 
@@ -36,6 +42,7 @@ export async function calendarGetEvents() {
  * @returns OK response??
  */
 export async function calendarCreateNewEvent(newEvent) {
+  // const user = useContext(UserContext);
   console.log('(CAPIHandling) Beginning of CalendarCreateNewEvent');
   try {
     console.log('New Event:' + JSON.stringify(newEvent, undefined, 2));
@@ -46,6 +53,7 @@ export async function calendarCreateNewEvent(newEvent) {
         headers: {
           Accept: 'application/json',
           'Content-Type': 'application/json',
+          // Authorization: `Bearer ${token}`,
         },
         body: newEvent,
       },
