+++ MobileUI/screens/LoginScreen.js
@@ -1,184 +1,25 @@
 import React, {useState, useContext} from 'react';
-import {Text, SafeAreaView, StyleSheet, View, TextInput} from 'react-native';
+import {
+  Text,
+  SafeAreaView,
+  StyleSheet,
+  View,
+  TextInput,
+  Alert,
+} from 'react-native';
 import Colors from '../assets/styles/colors';
 import {GreyPillButton} from '../assets/components/CustomButtons';
 import LogStateContext from '../contexts/LoginState';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import UserContext from '../contexts/User';
 import {loginAPICall} from '../API/LoginAPIHandler';
-
-const styles = StyleSheet.create({
-  screenContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: Colors.DD_CREAM,
-    color: Colors.DD_RED_2,
-  },
-  defaultScreentext: {
-    fontSize: 25,
-    fontWeight: '500',
-    color: Colors.DD_RED_3,
-    textAlign: 'center',
-    marginBottom: 46,
-  },
-  userName: {
-    height: 51,
-    width: 260,
-    marginLeft: 57,
-    marginRight: 60,
-    marginTop: 62,
-    backgroundColor: Colors.DD_RED_2,
-    paddingLeft: 17,
-    color: Colors.DD_LIGHT_GRAY,
-  },
-  userPassword: {
-    height: 51,
-    width: 260,
-    marginLeft: 56,
-    marginRight: 60,
-    marginTop: 37,
-    backgroundColor: Colors.DD_RED_2,
-    paddingLeft: 17,
-    color: Colors.DD_LIGHT_GRAY,
-  },
-});
-
-const setStringUsername = async (username, value) => {
-  try {
-    await AsyncStorage.setItem(`${username}`, value);
-  } catch (e) {
-    // save error
-    console.log('error : ' + e);
-    throw e;
-  }
-
-  // console.log('Set username in async storage is done.');
-};
-
-export const storeUserLoginInfo = async (name, password, postResult) => {
-  let loginInfo;
-  if (postResult.status !== '') {
-    loginInfo = {
-      password: password,
-    };
-  } else {
-    loginInfo = {
-      token: postResult.token,
-      expiration: postResult.expiration,
-      password: password,
-    };
-  }
-
-  await setStringUsername(name, JSON.stringify(loginInfo));
-};
-
-/*
- * Clears all contents in AsyncStorage
- */
-export const clearAll = async () => {
-  try {
-    await AsyncStorage.clear();
-  } catch (e) {
-    // clear error
-    console.log('!!!! Error with clearing! > ' + e);
-    throw e;
-  }
-
-  console.log('LoginScreen: Done clearing async storage');
-};
-
-/*
- * Gets the value stored with given username
- */
-const getUsernameValue = async username => {
-  try {
-    const value = await AsyncStorage.getItem(username);
-    if (value !== null) {
-      // console.log('getUsername: value=' + value);
-      return JSON.parse(value);
-    }
-  } catch (e) {
-    // read error
-    console.log('getUsername error: ' + e);
-    throw e;
-  }
-  // console.log('Done getting usesrname');
-  return null;
-};
-
-export const getAllKeys = async () => {
-  let keys = [];
-  try {
-    keys = await AsyncStorage.getAllKeys();
-    if (keys !== null) {
-      console.log('keys: ' + keys);
-      return keys;
-    }
-  } catch (e) {
-    // read key error
-    console.log('getAllKeys error: ' + e);
-    throw e;
-  }
-
-  return keys;
-  // example console.log result:
-  // ['@MyApp_user', '@MyApp_key']
-};
-
-export async function isStorageEmpty() {
-  // console.log(
-  //   'isStorageEmpty: Checking if storage is empty. Call getAllKeys()',
-  // );
-  let k = await getAllKeys();
-  if (k.length > 0) {
-    // console.log('isStorageEmpty: No');
-    return false;
-  } else {
-    // console.log('isStorageEmpty: Yes');
-    return true;
-  }
-}
-
-// export const CheckIsTokenExpired = async () => {
-export async function checkIsTokenExpired(name) {
-  // console.log('checkIsTExp - step 2.1: Call isStorageEmpty');
-
-  let emptyStorage = await isStorageEmpty();
-  if (!emptyStorage) {
-    //There are things that exist in storage
-    // console.log(
-    //   'checkIsTExp - step 2.3: There are many!! Parse expiration and current date',
-    // );
-
-    let user = await getUsernameValue(name);
-    // console.log('checkIsTExp - step 2.4: user result=' + user);
-    var exp = Date.parse(user.expiration);
-    var d1 = new Date();
-    var d = Date.parse(d1);
-
-    // console.log(
-    //   'checkIsTExp - step 2.5: Check if current date is past expiration',
-    // );
-    if (d >= exp) {
-      // console.log('checkIsTExp - Token is expired.');
-      return true;
-    } else {
-      // console.log('checkIsTExp - Token is not expired');
-      return false;
-    }
-  } else {
-    // There is nothing stored. Need to login. return true
-    // console.log('checkIsTExp - step 2.6: nothing is stored, return true');
-    return true;
-  }
-}
+import {storeUserLoginInfo, clearAll} from '../miscHelpers/AsyncStorageMethods';
 
 const LoginScreen = ({navigation}) => {
   const [name, setName] = useState('');
   const [password, setPassword] = useState('');
   const [isLoading, setIsLoading] = useState(false);
-  const [err, setErr] = useState('');
+  // const [err, setErr] = useState('');
   const {isLoggedIn, setIsLoggedIn} = useContext(LogStateContext);
   const user = useContext(UserContext);
 
@@ -202,11 +43,19 @@ const LoginScreen = ({navigation}) => {
   });
 
   const handleLogInButton = async () => {
-    // console.log('setisloading to true');
+    console.log('(1)setisloading to true');
     setIsLoading(true);
-
-    const response = await loginAPICall(name, password);
-    // console.log('await response');
+    let response = '';
+    try {
+      console.log('(2)calling loginAPICall');
+      response = await loginAPICall(name, password);
+      console.log('(3)line after call');
+    } catch (err) {
+      console.log('set is loading false. Send an alert for this eror: ' + err);
+      Alert.alert('Username or password is not correct, try again');
+      // throw err;
+    }
+    console.log('(4)await response');
     const result1 = await response.json();
     // console.log(result1);
 
@@ -269,10 +118,46 @@ const LoginScreen = ({navigation}) => {
         {!!isLoading && (
           <Text style={styles.defaultScreentext}>Loading...</Text>
         )}
-        {{err} && <Text style={styles.defaultScreentext}>{err}</Text>}
       </View>
     </SafeAreaView>
   );
 };
+
+const styles = StyleSheet.create({
+  screenContainer: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: Colors.DD_CREAM,
+    color: Colors.DD_RED_2,
+  },
+  defaultScreentext: {
+    fontSize: 25,
+    fontWeight: '500',
+    color: Colors.DD_RED_3,
+    textAlign: 'center',
+    marginBottom: 46,
+  },
+  userName: {
+    height: 51,
+    width: 260,
+    marginLeft: 57,
+    marginRight: 60,
+    marginTop: 62,
+    backgroundColor: Colors.DD_RED_2,
+    paddingLeft: 17,
+    color: Colors.DD_LIGHT_GRAY,
+  },
+  userPassword: {
+    height: 51,
+    width: 260,
+    marginLeft: 56,
+    marginRight: 60,
+    marginTop: 37,
+    backgroundColor: Colors.DD_RED_2,
+    paddingLeft: 17,
+    color: Colors.DD_LIGHT_GRAY,
+  },
+});
 
 export default LoginScreen;
